+++ src/main/kotlin/io/github/malczuuu/gradle/nullmarked/PackagePattern.kt
@@ -30,8 +30,6 @@ internal class PackagePattern private constructor(private val regex: Regex) {
 
   companion object {
 
-    private val VALID_IDENTIFIER = Regex("""(\.\.)?[\w*]+((\.|\.\.)[\w*]+)*(\.\.)?|\.\.""")
-
     fun of(identifier: String): PackagePattern {
       require(VALID_IDENTIFIER.matches(identifier)) { "Invalid package identifier '$identifier' [nullmarked plugin]." }
       return PackagePattern(Regex(toRegex(identifier)))